Organize Vercel setup modules
