+++ src/setup/vercel/index.ts
@@ -1,0 +1,3 @@
+export { applyVercelProfile } from "./profile.js";
+export type { VercelSetupCli, VercelSetupServices } from "./setup.js";
+export { setupVercel } from "./setup.js";
